Chore: Clean up root module namespace
